\name{makeBrick}
\docType{methods}
\alias{makeBrick}
\alias{makeBrick-methods}
\alias{makeBrick,RasterLayer-method}
\title{ Create a RasterBrick object}
\description{
 RasterBrick can be created from several objects.
}
\section{Methods}{
\describe{

\item{object = "RasterLayer"}{ make a RasterBrick from one or more RasterLayers }
}}
\keyword{methods}
\keyword{spatial}
