\name{Rcpp-class}

\docType{class}


\alias{SpatExtent}
\alias{SpatPoly}
\alias{SpatPolyPart}
\alias{SpatPolygons}
\alias{SpatExtent}
\alias{SpatPoly-class}
\alias{SpatPolyPart-class}
\alias{SpatPolygons-class}
\alias{Rcpp_SpatExtent-class}
\alias{Rcpp_SpatPolygons-class}
\alias{Rcpp_SpatPoly-class}
\alias{Rcpp_SpatPolyPart-class}

\title{ Rcpp classes}

\description{
These classes are for internal use only
}


\keyword{classes}
\keyword{spatial}
