\name{reclass}

\alias{reclass}
\alias{reclass,Raster-method}

\title{Reclass}

\description{
Deprecated. Use \code{\link{reclassify}} instead.
}

\usage{
reclass(x, rcl, ...)
}

\arguments{
  \item{x}{See \code{\link{reclassify}}}
  \item{rcl}{See \code{\link{reclassify}}}
  \item{...}{See \code{\link{reclassify}}} 
}

\seealso{ \code{\link{reclassify} } } 
