\name{rotated}

\alias{rotated}

  
\title{Do the raster cells have a rotation?}

\description{
Do the raster cells have a rotation?
}

\usage{
rotated(x)
}

\arguments{
\item{x}{A Raster* object} 
}
 

\value{
Logical value
}
  


\seealso{ \code{\link{rectify}}}

\examples{
r <- raster()
rotated(r)
}

\keyword{spatial}
