\name{Summary}

\docType{methods}
\alias{summary}
\alias{summary,RasterLayer-method}
\alias{summary,RasterStackBrick-method}
\alias{show,RasterSummary-method}


\title{Summary}

\description{
Summarize a Raster* object. Only useful for Raster* object that have their values in memory. See \link{cellStats} for other cases.
}

\value{A RasterSummary object}


\keyword{methods}
\keyword{spatial}
