\name{tile}

\alias{readTile}
\alias{tileSize}


\title{ tiles / bocks }

\description{
under development
}

\usage{
tileSize(x)
readTile(x, bs, b)
}


\arguments{
  \item{x}{Raster object}
  \item{bs}{ block/tile size }
  \item{b}{ number of blocks }
 }
 
\value{
numeric or data.frame
} 

\author{ Robert J. Hijmans }


\keyword{ spatial }
