\name{click}
\alias{click}
\alias{clickBbox}

\title{ Click on map }
\description{
  Click on a map to get values, coordinates or a bounding box
}
\usage{
click(object, n=1, xy=FALSE, type = "n", ...)
clickBbox(show = TRUE, col = "red") 
}

\arguments{
  \item{object}{A RasterLayer}
  \item{n}{number of clicks on the map}
  \item{xy}{show xy coordinates}
  \item{type}{see under ?locator}
  \item{show}{logical. If \code{TRUE}, the bounding box will be drawn on the map}  
  \item{col}{the color of the bounding box lines}    
 }
 
\details{
forthcoming
}


\author{ Robert J. Hijmans }

\examples{ 
 }

\keyword{ spatial }
