\name{Cluster}

\alias{setCluster}

\title{Use a cluster or multi-core cpu}

\description{
... Under development ...
}

\usage{
setCluster(n, type)
}

\arguments{
\item{n}{Integer. The number of nodes to be used (optional)}
\item{type}{Character. The cluster type to be used (optional)}
}

\value{
None. The side effect is to set global options.
}

\author{Matteo Mattiuzzi and Robert J. Hijmans}

\keyword{ spatial }

