\name{density}

\alias{density}
\alias{density,RasterLayer-method}

\docType{methods}

\title{density}

\description{
Create a density plots of values in a RasterLayer
}

\usage{
density(x, ...)
}

\arguments{
\item{x}{RasterLayer object}
\item{...}{Additional arguments passed to plot}
}

\value{
A density plot. 
}

\author{Robert J. Hijmans}


\keyword{spatial}
