\name{erase}

\docType{methods}

\alias{erase}
\alias{erase,SpatialPolygons,SpatialPolygons-method}

\title{ 
Erase parts of a Spatial* object
}

\description{
Erase parts of a Spatial* objects with another Spatial* object
}

\usage{
\S4method{erase}{SpatialPolygons,SpatialPolygons}(x, y, ...)
}

\arguments{
  \item{x}{Spatial* object}
  \item{y}{Spatial* object}
  \item{...}{Additional arguments (none)}  
}

\value{
Spatial*
}

\author{Robert J. Hijmans}


\keyword{methods}
\keyword{spatial}

