\name{worldFile}

\alias{worldFile}

\title{Header files}

\description{
Create a (ESRI) "World file", that can be used to georeference raster (image) data.
}

\usage{
worldFile(raster, extension=".wld") 
}

\arguments{
  \item{raster}{ RasterLayer object }
  \item{extension}{ extension of header file}
}


\author{Robert J. Hijmans}


\keyword{ spatial }
