\name{getData}
\alias{getData}
\alias{ccodes}

\title{Get geographic data }

\description{
This function has been deprecated. See the geodata package for access to more recent versions of the data available through this function, and for many other datasets.
}

\usage{
getData(name, download=TRUE, path="", ...)
ccodes()
}

\arguments{
  \item{name}{Data set name}
  \item{download}{logical}
  \item{path}{character}
  \item{...}{Additional arguments}   
}
 
\value{
Raster* or Spatial*
} 

\keyword{ spatial }
