\name{inifile}
\alias{readIniFile}

\title{Miscellaneous functions}
\description{
	Read ini files
}
\usage{
readIniFile(filename)
}
\arguments{
  \item{filename}{filename of the .ini file}
}

\value{
  readIniFile reads "ini-files". I.e. text file databases that are orginized in [sections] and "name = value" pairs.
}

\author{Robert J. Hijmans}

\keyword{file}
