\name{intersect}

\docType{methods}

\alias{intersect,Extent,Extent-method}

\title{ 
Intersect Extent
}

\description{
Intersect two Extent objects. O
}

\usage{
\S4method{intersect}{Extent,Extent}(x, y)
}

\arguments{
  \item{x}{Extent}
  \item{y}{Extent}
} 

\note{
use \code{\link{crop}} if you want to intersect a Spatial* object with an Extent object.
}

\value{
Extent
}

\author{Robert J. Hijmans}

\seealso{
\code{\link{unionExtent}, \link{intersectExtent}}}
}


\keyword{methods}
\keyword{spatial}

