\name{KML}

\alias{KML}

\title{KML}

\description{
Export raster to a KML file
}

\usage{
KML(raster, filename, col=rainbow(255), maxdim=1000) 
}

\arguments{
\item{raster}{RasterLayer}
\item{filename}{output filename}
\item{col}{color scheme to be used}
\item{maxdim}{sampling scheme}
}

\details{
}

\value{
None. But two files are written to disk. 
}

\author{adapted for the raster package by Robert J. Hijmans; based on functions in the maptools package by Duncan Golicher, David Forrest and Roger Bivand }

\examples{
}

\keyword{spatial}
