\name{Logic-methods}
\docType{methods}
\alias{Logic-methods}
\alias{Logic,RasterLayer,RasterLayer-method}
\alias{is.na,RasterLayer-method}

\title{ Methods for logical functions for RasterLayers' }
\description{
	The following boolean operators are available: '&', '|'
	is.na(object) can also be called. 
}

\keyword{methods}
\keyword{math}
