\name{Logic-methods}
\docType{methods}
\alias{Logic-methods}
\alias{Logic,RasterLayer,RasterLayer-method}
\alias{is.na,RasterLayer-method}
\alias{is.nan,RasterLayer-method}
\alias{is.finite,RasterLayer-method}
\alias{is.infinite,RasterLayer-method}

\title{ Methods for logical functions for RasterLayers' }
\description{
	The following boolean operators are available: '&', '|'
	These are also available (with object being a RasterLayer, and the result a RasterLayer with logical values:
	is.na(object) 
	is.nan(object)
	is.finite(object)
	is.infinite(object)
}

\keyword{methods}
\keyword{math}
