\alias{adjacency}
\alias{layerNames}
\alias{count}
\alias{layerNames,Raster-method}
\alias{count,Raster-method}

\name{Obsolete}

\title{Obsolete (depracated) functions}

\description{
These functions are obsolete. 

adjacency has been replaced by \code{\link{adjacent}}

layerNames has been replaced by \code{\link{names}}

count has been replaced by \code{\link{freq}}
}

\usage{
adjacency(x, ...) 
count(x, ...)
layerNames(x, ...)
}

\arguments{
\item{x}{Raster* object}
\item{...}{additional arguments}
}


\keyword{spatial}

