
\alias{setOptions}
\alias{showOptions}
\alias{clearOptions}
\alias{saveOptions}

\alias{intersectExtent}
\alias{unionExtent}
\alias{count}


\name{Obsolete}

  
\title{Obsolete (depracated) functions}

\description{
Functions setOptions, showOptions, saveOptions, and clearOptions have been replaced by \code{\link{rasterOptions}}

intersectExtent has been replaced by \code{\link{intersect}}
unionExtent has been replaced by \code{\link{union}}

count has been replaced by freq( , , value=  )
}

\usage{
setOptions(...)
showOptions()
saveOptions()
clearOptions()
intersectExtent(...)
unionExtent(...)
count(...)
}

\arguments{
\item{...}{arguments}
}

\seealso{ \code{\link{rasterOptions}}, \code{\link{intersect}}, \code{\link{union}} }

\keyword{spatial}

