\name{obsolete}

\alias{polygonValues}
\alias{lineValues}
\alias{xyValues}

\alias{pointsToRaster}
\alias{linesToRaster}
\alias{polygonsToRaster}


\title{ Obsolete functions }

\description{
Obsolete functions. Use \code{\link{extract}} or \code{\link{rasterize}} instead.
}

\usage{
polygonValues(...)
lineValues(...)
xyValues(...)
pointsToRaster(...)
linesToRaster(...)
polygonsToRaster(...)
}

\arguments{
\item{...}{arguments}
}

\value{
An error message
}

\seealso{ \code{\link[raster]{extract}, \link[raster]{rasterize}} }


\keyword{methods}
