\name{obsolete}

\alias{polygonValues}
\alias{lineValues}
\alias{xyValues}

\title{ Obsolete functions }

\description{
Obsolete functions. Use \code{\link[raster]{extract}} instead.
}

\usage{
polygonValues(...)
lineValues(...)
xyValues(...)
}

\arguments{
\item{...}{arguments}
}

\value{
An error message
}

\seealso{ \code{\link[raster]{extract} } }


\keyword{methods}
