
\alias{adjacency}

\name{Obsolete}

\title{Obsolete (depracated) functions}

\description{
adjacency has been replaced by \code{\link{adjacent}}
}

\usage{
adjacency(x, fromCells, toCells, directions) 
}

\arguments{
\item{x}{Raster* object}
\item{fromCells}{integer}
\item{toCells}{integer}
\item{directions}{integer}
}

\seealso{ \code{\link{adjacent}} }

\keyword{spatial}

