\name{polygonValues}

\docType{methods}

\alias{polygonValues}
\alias{lineValues}

\title{ Extract values from a RasterLayer by polygons or lines}

\description{
Obsolete functions. Use \code{\link[raster]{extract}} in stead.
}

\usage{
polygonValues(p, x, ...)
lineValues(lns, x, ...)
}

\arguments{
\item{p}{SpatialPolygons* object}
\item{lns}{SpatialLines* object}
\item{x}{Raster* object}
\item{...}{additional arguments}
}

\value{
A list
}

\seealso{ \code{\link[raster]{extract} } }

\author{Robert J. Hijmans }


\keyword{methods}
\keyword{spatial}
