\name{raster-package}
\alias{raster-package}
\alias{raster}
\docType{package}
\title{
Analysis and modeling of geographic (spatial) data.
}
\description{
This packages implements functionality to use and manipulate 'raster' (or 'grid') data. That is geographic data represented as a matrix).
It does this by adding new functions and by pulling together existing R functions in a single package that aims at being easy to use. 
}
\details{
\tabular{ll}{
Package: \tab raster\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2008-09-11\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Overview here.....

}
\author{
Robert J. Hijmans \email{r.hijmans@gmail.com}
Jacob van Etten

Maintainer: Robert J. Hijmans \email{r.hijmans@gmail.com}
}
\references{

}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link[<pkg>:<sp>-package]{<sp>}}
}
