\name{raster-package}
\alias{raster-package}
\alias{raster}
\docType{package}
\title{
Analysis and modeling of geographic (spatial) data.
}
\description{
This packages implements functionality to use and manipulate 'raster' (or 'grid') data (geographic data that can be represented as a matrix).
It does this by adding new functions and by pulling together existing R functions in a single package that aims at being easy to use. 
}
\details{
\tabular{ll}{
Package: \tab raster\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2008-09-11\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
This package should be particularly useful when using rather large datasets, as it can
use datasources that are stored on disk without loading them into memory. 

}
\author{
Robert J. Hijmans and Jacob van Etten

Maintainer: Robert J. Hijmans \email{r.hijmans@gmail.com}
}
\keyword{ package }
\keyword{ spatial }
