\name{raster-package}
\alias{raster-package}
\alias{raster}
\docType{package}
\title{
Analysis and modeling of geographic (spatial) raster data.
}
\description{
This packages provides classes and functions to manipulate spatial data in 'raster' (also called 'grid') form. 

}
\details{
\tabular{ll}{
Package: \tab raster\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2009-01-11\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
This package should be particularly useful when using rather large datasets, as it can
use datasources that are stored on disk without loading them into memory. 

}
\author{
Robert J. Hijmans and Jacob van Etten

Maintainer: Robert J. Hijmans \email{r.hijmans@gmail.com}
}
\keyword{ package }
\keyword{ spatial }
