\name{raster.utilities}
\alias{raster.compare}
\alias{create.CRS}
  
\title{ Compare rasters }
\description{
  raster.compare evaluates whether a list of rasters has the same origin, resolution, number of rows and columns, and projection (or a subset thereof).
}

\usage{
raster.compare(rasters, origin=TRUE, resolution=TRUE, rowcol=TRUE, projection=TRUE, slack=0.01, stopiffalse=TRUE) 
}

\arguments{
  \item{rasters} {}
  \item{origin} {}
  \item{resolution}{}
  \item{rowcol}{}
  \item{projection}{}
  \item{slack} {}
  \item{stopiffalse}{}
 }
  
\value{
}

\author{Robert J. Hijmans  \email{r.hijmans@gmail.com}}
\examples{
}
\keyword{ spatial }
