\name{Rcpp-class}

\docType{class}


\alias{SpExtent}
\alias{SpPoly}
\alias{SpPolyPart}
\alias{SpPolygons}
\alias{SpExtent}
\alias{SpPoly-class}
\alias{SpPolyPart-class}
\alias{SpPolygons-class}
\alias{Rcpp_SpExtent-class}
\alias{Rcpp_SpPolygons-class}
\alias{Rcpp_SpPoly-class}
\alias{Rcpp_SpPolyPart-class}

\title{ Rcpp classes}

\description{
These classes are for internal use only
}


\keyword{classes}
\keyword{spatial}
