\name{Slope and aspect}

\alias{slopeAspect}

\title{Slope and aspect}

\description{
DEPRACATED. Use \code{\link{terrain}} instead. 
}

\usage{
slopeAspect(dem, filename='', out=c('slope', 'aspect'), unit='radians', 
                 neighbors=8, flatAspect, ...)  
}

\arguments{
  \item{dem}{DEPRACATED}
  \item{filename}{DEPRACATED}
  \item{out}{DEPRACATED}
  \item{unit}{DEPRACATED}
  \item{neighbors}{DEPRACATED}
  \item{flatAspect}{DEPRACATED}
  \item{...}{DEPRACATED}
}

\seealso{ \code{\link{terrain}} }


\keyword{spatial}
