\name{Summarize}
\docType{methods}
\alias{summary}
\alias{summary,RasterLayer-method}
\alias{summary,RasterStack-method}
\alias{show,RasterLayerSummary-method}


\title{ Summary}
\description{
 Summarize a RasterLayer or RasterStack
}
\section{Methods}{
\describe{

-

}}
\keyword{methods}
\keyword{spatial}
