\name{Summarize}

\docType{methods}
\alias{summary}
\alias{summary,RasterLayer-method}
\alias{summary,RasterStack-method}
\alias{show,RasterLayerSummary-method}


\title{ Summarize}
\description{
 Summarize a RasterLayer or RasterStack
}
\section{Methods}{
\describe{

Creates an object of class "RasterLayerSummary". Not particularly useful.

}}
\keyword{methods}
\keyword{spatial}
