\name{Summary}

\docType{methods}

\alias{summary,Raster-method}
\alias{summary,RasterStack-method}
\alias{summary,RasterBrick-method}
\alias{show,RasterSummary-method}


\title{Summary}

\description{
Summarize a Raster* object. A sample is used for very large files. 
}




\details{
\code{ summary(x, maxsamp=5000, ...) }
You can set the sample size with \code{maxsamp} 
}

\value{A RasterSummary object}


\seealso{ \code{\link{cellStats}}  }


\keyword{methods}
\keyword{spatial}
