\name{symdif}

\docType{methods}

\alias{symdif}
\alias{symdif,SpatialPolygons,SpatialPolygons-method}

\title{ 
Symetrical difference
}

\description{
Symetrical difference of SpatialPolygons* objects
}

\usage{
\S4method{symdif}{SpatialPolygons,SpatialPolygons}(x, y, ...)
}

\arguments{
  \item{x}{SpatialPolygons* object}
  \item{y}{SpatialPolygons* object}
  \item{...}{Additional SpatialPolygons* object(s)}  
}

\value{
SpatialPolygons*
}

\author{Robert J. Hijmans}

\seealso{
\code{\link{erase}} 
}


\keyword{methods}
\keyword{spatial}

