\name{writeStack}

\alias{writeStack}

\title{Write values of a RasterStack to a file}

\description{
xxx
}

\usage{
  writeStack(rstack, overwrite=FALSE)
}

\arguments{
  \item{rstack}{RasterStack object} 
  \item{overwrite}{Logical. If \code{TRUE} existing files are overwritten}
  
}

\details{
xxx
}


\author{Robert J. Hijmans}

 
\keyword{ spatial }
