\name{writeStack}

\alias{writeStack}

\title{Write RasterStack}

\description{

NOT IMPLEMENTED YET

values of a RasterStack to a multi-band RasterFile
}

\usage{
  writeStack(rstack, overwrite=FALSE)
}

\arguments{
  \item{rstack}{RasterStack object} 
  \item{overwrite}{Logical. If \code{TRUE} existing files are overwritten}
  
}

\details{
xxx
}


\author{Robert J. Hijmans}

 
\keyword{ spatial }
